backend implementation

+++ backend/app/routers/licenses.py
@@ -5,9 +5,11 @@ from fastapi import HTTPException, Depends, APIRouter
 
 from app.deps.authorization_deps import GroupAuthorization
 from app.keycloak_auth import get_current_user, get_user
-from app.models.groups import GroupOut, GroupDB
-from app.models.users import UserDB
+
+from backend.app.deps.authorization_deps import Authorization
+from backend.app.models.users import UserDB
 from backend.app.models.licenses import LicenseOut, LicenseIn, LicenseDB, LicenseBase
+from backend.app.routers.authentication import get_admin, get_admin_mode
 
 router = APIRouter()
 
@@ -22,24 +24,27 @@ async def save_license(
     return license_db.dict()
 
 
-@router.get("/{license_id}", response_model=GroupOut)
-async def get_license(
-    license_id: str,
-    allow: bool = Depends(GroupAuthorization("viewer")),
-):
+@router.get("/{license_id}", response_model=LicenseOut)
+async def get_license(license_id: str):
     if (license := await LicenseDB.get(PydanticObjectId(license_id))) is not None:
         return license.dict()
-    raise HTTPException(status_code=404, detail=f"Group {license_id} not found")
+    raise HTTPException(status_code=404, detail=f"License {license_id} not found")
 
 
-@router.put("/{license_id}", response_model=GroupOut)
+@router.put("/{license_id}", response_model=LicenseOut)
 async def edit_license(
     license_id: str,
     license_info: LicenseBase,
     user_id=Depends(get_user),
-    allow: bool = Depends(GroupAuthorization("editor")),
+    admin=Depends(get_admin),
+    admin_mode: bool = Depends(get_admin_mode),
 ):
-    if (license := await GroupDB.get(PydanticObjectId(license_id))) is not None:
+    if (license := await LicenseDB.get(PydanticObjectId(license_id))) is not None:
+        if license.creator != user_id and not (admin and admin_mode):
+            raise HTTPException(
+                status_code=403,
+                detail=f"User {user_id} doesn't have permission to edit license {license_id}",
+            )
         license_dict = dict(license_info) if license_info is not None else {}
 
         if len(license_dict["name"]) == 0 or len(license_dict["holders"]) == 0:
@@ -49,27 +54,34 @@ async def edit_license(
             )
             return
 
-        user = await UserDB.find_one(UserDB.email == user_id)
-        license.creator = user.dict()
         license.modified = datetime.utcnow()
         license.holders = license_dict["holders"]
-        license.type = license_dict["text"]
+        license.type = license_dict["type"]
+        license.text = license_dict["text"]
         license.url = license_dict["url"]
         license.version = license_dict["version"]
         license.allow_download = license_dict["allow_download"]
+        license.name = license_dict["name"]
         license.replace()
 
         return license.dict()
     raise HTTPException(status_code=404, detail=f"License {license_id} not found")
 
 
-@router.delete("/{license_id}", response_model=GroupOut)
+@router.delete("/{license_id}", response_model=LicenseOut)
 async def delete_license(
     license_id: str,
-    allow: bool = Depends(GroupAuthorization("owner")),
+    user_id=Depends(get_user),
+    admin=Depends(get_admin),
+    admin_mode: bool = Depends(get_admin_mode),
 ):
     if (license := await LicenseDB.get(PydanticObjectId(license_id))) is not None:
+        if license.creator != user_id and not (admin and admin_mode):
+            raise HTTPException(
+                status_code=403,
+                detail=f"User {user_id} doesn't have permission to delete license {license_id}",
+            )
         await license.delete()
         return license.dict()  # TODO: Do we need to return what we just deleted?
     else:
-        raise HTTPException(status_code=404, detail=f"Dataset {license_id} not found")
+        raise HTTPException(status_code=404, detail=f"License {license_id} not found")
